Apply feedback.
