+++ alliander_robotics/alliander_robotiq/src/alliander_robotiq/CMakeLists.txt
@@ -9,6 +9,9 @@ project(alliander_robotiq)
 find_package(ament_cmake REQUIRED)
 find_package(ament_cmake_python REQUIRED)
 
+# Python package:
+ament_python_install_package(${PROJECT_NAME})
+
 # Python executables:
 install(
   DIRECTORY src_py/
