+++ alliander_robotics/alliander_robotiq/src/alliander_robotiq/alliander_robotiq/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: Alliander N. V.
+#
+# SPDX-License-Identifier: Apache-2.0
